feat: prepares HTTP handler to serve LPS v1 and v2
